fix data type error

+++ include/StatTimer.h
@@ -57,24 +57,21 @@
 
 #if defined(_WIN32)
 #if !defined (_WINSOCK2API_) && !defined(_WINSOCKAPI_)
-struct timeval
-{
-	long    tv_sec;         /* seconds */
-	long    tv_usec;        /* and microseconds */
+struct timeval {
+  long    tv_sec;         /* seconds */
+  long    tv_usec;        /* and microseconds */
 };
 #endif
 
-inline static int gettimeofday(struct timeval *tv, void *tz)
-{
-	union
-	{
-		long long ns100;
-		FILETIME    t;
-	} now;
-	GetSystemTimeAsFileTime(&now.t);
-	tv->tv_usec = long((now.ns100 / 10LL) % 1000000LL);
-	tv->tv_sec = long((now.ns100 - 116444736000000000LL) / 10000000LL);
-	return 0;
+inline static int gettimeofday(struct timeval* tv, void* tz) {
+  union {
+    long long ns100;
+    FILETIME    t;
+  } now;
+  GetSystemTimeAsFileTime(&now.t);
+  tv->tv_usec = static_cast<long>((now.ns100 / 10LL) % 1000000LL);
+  tv->tv_sec = static_cast<long>((now.ns100 - 116444736000000000LL) / 10000000LL);
+  return 0;
 }
 #undef HAS_CLOCK_GETTIME
 
@@ -86,57 +83,55 @@ inline static int gettimeofday(struct timeval *tv, void *tz)
 
 /// Class to abstract socket communications in a cross platform manner.
 /// This class is designed
-class CStatTimer
-{
-public:
-	CStatTimer()
-	{
-	};
+class CStatTimer {
+ public:
+  CStatTimer() {
+  };
 
-	~CStatTimer()
-	{
-	};
+  ~CStatTimer() {
+  };
 
-	void Initialize()
-	{
-		memset(&m_startTime, 0, sizeof(struct timeval));
-		memset(&m_endTime, 0, sizeof(struct timeval));
-	};
+  void Initialize() {
+    memset(&m_startTime, 0, sizeof(struct timeval));
+    memset(&m_endTime, 0, sizeof(struct timeval));
+  };
 
-	struct timeval GetStartTime() { return m_startTime; };
-	void SetStartTime() { GET_CLOCK_COUNT(&m_startTime); };
+  struct timeval GetStartTime() { return m_startTime; };
+  void SetStartTime() {
+    GET_CLOCK_COUNT(&m_startTime);
+  };
 
-	struct timeval GetEndTime() { return m_endTime; };
-	void SetEndTime() { GET_CLOCK_COUNT(&m_endTime); };
+  struct timeval GetEndTime() { return m_endTime; };
+  void SetEndTime() {
+    GET_CLOCK_COUNT(&m_endTime);
+  };
 
-	uint32_t GetMilliSeconds() { return (CalcTotalUSec() / MILLISECONDS_CONVERSION); };
-	uint64_t GetMicroSeconds() { return (CalcTotalUSec()); };
-	uint32_t GetSeconds() { return (CalcTotalUSec() / MICROSECONDS_CONVERSION); };
+  uint32_t GetMilliSeconds() { return (CalcTotalUSec() / MILLISECONDS_CONVERSION); };
+  uint64_t GetMicroSeconds() { return (CalcTotalUSec()); };
+  uint32_t GetSeconds() { return (CalcTotalUSec() / MICROSECONDS_CONVERSION); };
 
-	static uint64_t GetCurrentTime()
-	{
+  static uint64_t GetCurrentTime() {
 #if HAS_CLOCK_GETTIME
-		struct timespec  tim;
-		clock_gettime(CLOCK_REALTIME, &tim);
-		return (uint64_t)(tim.tv_sec * 1000000000LL + tim.tv_nsec);
+    struct timespec  tim;
+    clock_gettime(CLOCK_REALTIME, &tim);
+    return (uint64_t)(tim.tv_sec * 1000000000LL + tim.tv_nsec);
 #else
-		struct timeval timeofday;
-		gettimeofday(&timeofday, NULL);
-		return (uint64_t)(timeofday.tv_sec * 1000000000LL + timeofday.tv_usec * 1000);
+    struct timeval timeofday;
+    gettimeofday(&timeofday, NULL);
+    return (uint64_t)(timeofday.tv_sec * 1000000000LL + timeofday.tv_usec * 1000);
 #endif
-	};
+  };
 
-private:
-	uint32_t CalcTotalUSec()
-	{
-		return (((m_endTime.tv_sec - m_startTime.tv_sec) * MICROSECONDS_CONVERSION) +
-				(m_endTime.tv_usec - m_startTime.tv_usec));
-	};
+ private:
+  uint64_t CalcTotalUSec() {
+    return (((m_endTime.tv_sec - m_startTime.tv_sec) * MICROSECONDS_CONVERSION) +
+            (m_endTime.tv_usec - m_startTime.tv_usec));
+  };
 
 
-private:
-	struct timeval  m_startTime;
-	struct timeval  m_endTime;
+ private:
+  struct timeval  m_startTime;
+  struct timeval  m_endTime;
 };
 
 #endif // __CSTATTIMER_H__
